ref: Move ourlog processing code into own file
